// apps/backend/src/appsync/API.ts
export enum VerificationStatus {
  UNCHECKED = 'UNCHECKED',
  PASS = 'PASS',
  FAIL = 'FAIL',
  PASS_MANUAL = 'PASS_MANUAL',
  FAIL_MANUAL = 'FAIL_MANUAL',
  REFER = 'REFER',
  WAIT = 'WAIT',
  ARCHIVED = 'ARCHIVED',
  INACTIVE = 'INACTIVE',
}
// ...
export enum EntityType {
  COMPANY = 'COMPANY',
  TRUST = 'TRUST',
  NOT_FOR_PROFIT = 'NOT_FOR_PROFIT',
  SELF_MANAGED_SUPER_FUND = 'SELF_MANAGED_SUPER_FUND',
  BPAY = 'BPAY',
}

export type AddressInput = {
  addressId?: string | null;
  contactName?: string | null;
  contactNumber?: string | null;
  address1: string;
  address2?: string | null;
  country: string;
  countryCode?: string | null;
  streetNumber?: string | null;
  streetName: string;
  unitNumber?: string | null;
  streetType?: string | null;
  city: string;
  state: string;
  stateCode?: string | null;
  postalCode: string;
};

export type ImageInput = {
  alt?: string | null;
  identityId?: string | null;
  key: string;
  level: S3UploadLevel;
  type?: string | null;
};

export enum S3UploadLevel {
  public = 'public',
  protected = 'protected',
  private = 'private',
}
// ...
export type UpdateEntityInput = {
  id: string;
  type?: EntityType | null;
  taxNumber?: string | null;
  name?: string | null;
  address?: AddressInput | null;
  logo?: ImageInput | null;
  paymentMethodId?: string | null;
  disbursementMethodId?: string | null;
  ubosCreated?: boolean | null;
  numUbosCreated?: number | null;
  verificationStatus?: VerificationStatus | null;
  pepCheckStatus?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  legalName?: string | null;
  contact?: EntityContactInput | null;
  gstRegistered?: boolean | null;
  role?: string | null;
  isFirm?: boolean | null;
  updatedAt?: string | null;
};

export type EntityContactInput = {
  firstName?: string | null;
  lastName?: string | null;
  email?: string | null;
  phone?: string | null;
};